refactor:#377 memberContext에서 기수를 비교하여 membershipType을 반환하도록 수정
